Add UserFollowerAPIView to get all the followers of the user
this API fetches all the follower of the currently logged in user and returns total number of followers and
all all the followe's profile details.

Add ProfileUpdateAPIView to update an existing user profile

ProfileUpdateAPIView API handles put and patch request to update the profile detials of an user
The API uses MultiPartParser to handle file upload efficientlyl.