Fix: core_apps.articles.filter foreignkey icontains bug and core_apps.articles.views.ArticleRetriveUpdateDestroyView old photo not being deleted in patch request.
The filter.ArticleFilter and articles.admin had directly placed the foreignkey in look up fields. fixed.
the ArticleRetriveUpdateDestroyView has still one bug to fix as the old photo is not being deleted after the user passes a patch request to update the banner, body image 1 or 2 images. should be fixed in next commit